refactor(frontend): split lib/programmatic.ts into types + editorial modules
PSEO-002: Extração mecânica — sem mudança funcional.

- lib/programmatic/types.ts: tipos, constantes, utilitários, fetchers
- lib/programmatic/editorial.ts: conteúdo editorial + FAQ generators
- lib/programmatic/index.ts: barrel re-export
- lib/programmatic.ts: re-export barrel (backward compat)
